Update HA & HACS

+++ custom_components/powercalc/helpers.py
@@ -1,12 +1,13 @@
+from decimal import Decimal
 from typing import Union
 
 from homeassistant.helpers.template import Template
 
 
-async def evaluate_power(power: Union[Template, float]) -> float:
+async def evaluate_power(power: Union[Template, Decimal]) -> Decimal:
     """When power is a template render it."""
 
     if isinstance(power, Template):
         return power.async_render()
 
-    return power
+    return Decimal(power)
